let Lntxbot have internal transfers correctly handled
Invoices fulfilled internally within lntxbot users (including within one same user) do not respond with any payment_hash or payment_preimage. Still, if the reply has succeeded status it means it has gone through. We can handle the situation by giving a checking_id based on the hash of the response (it will not be directly queriable, but at the same time it needs to be unique in the DB), assign a blank fee to value 0, and leave the preimage blank