refactor: split GET payments endpoint into public and private
also why do we check the status of the payment directly there twice,
once through `check_transaction_status` and once through
`fundingsource.check_status()` and should not do that on get? it
probably was a safety mechanism, but it bit us already in the past.

its also our most spammed endpoint on legend, it should not be that
expernsive to run!