feat: introduce `NotFoundError`
instead of asserting and returning Optional on get operation in crud, we
through and NotFoundError. in the context of an API request the FastApi
errorhandler will catch it and throw an 404 like we have for Invoice-
and PaymentError.

In non-api context we can simly try catch the notfound like so:
```python
try:
  url = await get_tinyurl("test")
except NotFoundError:
  # url does not exist do whatever
  pass
```