fix: oversight from my, setting the default for dates
how i did it does not work, it actually initialized the default date
once on the server start and keeps that. i was quite sure it will
recalculate now() on each model initialization, but that was a false
assumption.

the fix is to make it a required value, and initalizing it when creating
a model