fix: `/error` should not be a generic route
i think i mistakenly added it to generic.

if added it renders a weird errorpage without error, by removing it from
here, it will render the proper `error.html` for page not found.

`/error` is just a dynamic vue route and hard reloading on that `path`
should lead to a 404